If the system allows ipv4, then the system’s ipv4 dns server is given priority for querying, and if there are multiple dns results (IPV4 or IPV6), take one of the ipv4DNS results for tcping, and if it cannot be accessed, take an IPV6DNS record result for tcping Inquire. If the system allows ipv6, and ipv4 is not allowed, then the ipv6 dns server of the system is preferred for query, and if there are multiple dns results (IPV4 or IPV6), take one of the ipv6 DNS results for tcping. If the system allows ipv4 and ipv6 at the same time, then query according to the above rules that the system allows ipv4, otherwise try IPV6 query until a DNS result can be resolved normally. The above content has nothing to do with the two parameters -4 -6 in the script.
If the -4 parameter is used, only IPV4 is used for DNS query and only one IPV4DNS result in the result is used for tcping. If the -6 parameter is used, only IPV4 is used for DNS query and only one IPV6DNS result in the result is used for tcping. If you use the two parameters -4 -6, the above is the requirement.